Move binder shift removal into Term
This in anticipation of binder opening/closing depending on shift removal.
