+++ src/Hopper/Internal/Core/Term.hs
@@ -4,9 +4,12 @@ module  Hopper.Internal.Core.Term where
 
 import Hopper.Internal.Core.Literal
 import Hopper.Internal.Type.BinderInfo (BinderInfo)
-import Hopper.Utils.LocallyNameless (Bound(..), Depth(..), Slot(..))
+import Hopper.Utils.LocallyNameless (Bound(..), Slot(..), Depth(..), localDepth,
+                                     depthLevel)
 
+import Control.Lens ((+~), (%~), (^?), ix)
 import Data.Data
+import Data.Function ((&))
 import Data.Word (Word32)
 
 import qualified Data.Vector as V
@@ -38,6 +41,53 @@ data Term v =
         !(Term v) --- BODY
   deriving ({-Show1,Read1,Ord1,Eq1,-}Ord,Eq,Show,Read{-,Functor,Foldable-},Typeable{-,Traversable-})
 
+-- | Removes binder shifts from a @'Term' 'Bound'@
+removeBinderShifts :: Term Bound -> Term Bound
+removeBinderShifts = go $ repeat 0 -- NB: [] only works for well-formed ASTs,
+                                   --     where no de Bruijn variables are free.
+  where
+    -- A bump amount @b@ at index @k@ in 'BinderShifts' means that we have @b@
+    -- instances of @BinderLevelShiftUP k@ that we imagine we're tracking as we
+    -- move down the AST.
+    --
+    -- When we enter a new scope, we @(0:)@ our list, effectively turning each
+    -- tracked @BinderLevelShiftUP k@ to @BinderLevelShiftUP (k+1)@.
+    --
+    -- When we encounter a @BinderLevelShiftUP k@, we @(+1)@ the list at index
+    -- @k@.
+    --
+    -- When we encounter a variable of depth @d@, we bump the variable by the
+    -- sum of the @b@ amounts for all of the shifts the variable is affected by
+    -- -- namely, the first @d+1@ @b@ values in the list of shifts.
+    go :: [Word32] -> Term Bound -> Term Bound
+    go shifts (V var) = case var ^? localDepth.depthLevel of
+      Just depth ->
+        let adjustment = sum $ take (succ $ fromIntegral depth) shifts
+        in V $ var & localDepth.depthLevel +~ adjustment
+      Nothing ->
+        V var
+
+    go shifts (BinderLevelShiftUP k t) =
+      go (shifts & ix (fromIntegral k) %~ succ) t
+
+    go _shifts lit@(ELit _) = lit
+
+    go shifts (Return ts) = Return $ go shifts <$> ts
+
+    go shifts (EnterThunk t) = EnterThunk $ go shifts t
+
+    go shifts (Delay t) = Delay $ go shifts t
+
+    go shifts (App ft ats) = App (go shifts ft) (go shifts <$> ats)
+
+    go shifts (PrimApp primId ts) = PrimApp primId $ go shifts <$> ts
+
+    -- In the 'Lam' and 'Let' cases, we enter a new scope:
+
+    go shifts (Lam infos t) = Lam infos $ go (0 : shifts) t
+
+    go shifts (Let infos rhs bod) = Let infos (go shifts rhs)
+                                              (go (0 : shifts) bod)
 
 --- NOTE: USE STE MONAD ONCE WE MIGRATE TO HSUM DESIGN
 --- this is kinda only for "inlining" on debruijin variables for now
@@ -75,13 +125,3 @@ substitute baseLevel initMapper initTerm = goSub 0 initMapper initTerm
     goSub _shift _mapper lit@(ELit _)  = Right lit
     goSub shift mapper (EnterThunk bod ) =  fmap EnterThunk $ goSub shift mapper bod
     goSub shift mapper (Delay bod ) = fmap Delay $ goSub shift mapper bod
-
-
-
-{-
-            _ _ (BinderLevelShiftUP _ _)
-            _ _ (ELit _)
-            _ _ (EnterThunk _)
-            _ _ (Delay _)
-
--}
